Implement a post_step command for the server
The current implementation of threads (threading is hard, yo!)
does not really allow to send 2 or more commands at once to the
server and read back their replies.

One such case is after a step command is done. At that time we'd
like to get all the variables and all the stacktraces. If we try
to get them at the same time, one thread will "overwrite" the other.

So I wrote a post_command method that will in a new thread first
get the variables and after that the stacktraces. By firing 2
separate events, one for variables, one for stacktraces, once the
thread is finished, the application's main event loop will handle
these 2 events seemingly at the same time as they get fired one
right after the other.