Reorganized settings window
Keeping widgets in a dict, indexed by their name, makes it easier to
implement methods like save_all and load_all. It also makes it easier to
add future settings.

Also added a new signal settings_changed_signal which includes a set of
names of the settings which were changed.