Simple ignoring of files and folders
Try to read in a gitignore file, find the folders
from it, and then ignore files from those folders.

This both speeds up and simplifies the searching
because we don't have to deal with a lot of noise
and tweak the fuzzy searches to always get the desired
results.

cpburnz/python-path-specification can be used to
have a better support for gitignore files.